yamitranscode: config encoder parameters
receive parameters from command line which used to config encoder.
