+++ tests/yamitranscode.cpp
@@ -23,7 +23,6 @@
 #include "config.h"
 #endif
 
-#include "yamitranscodehelp.h"
 #include "vppinputdecode.h"
 #include "vppinputoutput.h"
 #include "vppoutputencode.h"
@@ -36,10 +35,121 @@
 #include "VideoPostProcessHost.h"
 #include <stdio.h>
 #include <stdlib.h>
+#include <getopt.h>
 
 using namespace YamiMediaCodec;
 
-SharedPtr<VppInput> createInput(YamiEncodeParam& para, const SharedPtr<VADisplay>& display)
+static void print_help(const char* app)
+{
+    printf("%s <options>\n", app);
+    printf("   -i <source yuv filename> load YUV from a file or container\n");
+    printf("   -W <width> -H <height>\n");
+    printf("   -o <coded file> optional\n");
+    printf("   -b <bitrate: kbps> optional\n");
+    printf("   -f <frame rate> optional\n");
+    printf("   -c <codec: HEVC|AVC|VP8|JPEG>\n");
+    printf("   -s <fourcc: I420|NV12|IYUV|YV12>\n");
+    printf("   -N <number of frames to encode(camera default 50), useful for camera>\n");
+    printf("   --qp <initial qp> optional\n");
+    printf("   --rcmode <CBR|CQP> optional\n");
+}
+
+static VideoRateControl string_to_rc_mode(char *str)
+{
+    VideoRateControl rcMode;
+
+    if (!strcasecmp (str, "CBR"))
+        rcMode = RATE_CONTROL_CBR;
+    else if (!strcasecmp (str, "CQP"))
+        rcMode = RATE_CONTROL_CQP;
+    else {
+        printf("Unsupport  RC mode\n");
+        rcMode = RATE_CONTROL_NONE;
+    }
+    return rcMode;
+}
+
+static bool processCmdLine(int argc, char *argv[], TranscodeParams& para)
+{
+    char opt;
+    const struct option long_opts[] = {
+        {"help", no_argument, NULL, 'h' },
+        {"qp", required_argument, NULL, 0 },
+        {"rcmode", required_argument, NULL, 0 },
+        {NULL, no_argument, NULL, 0 }};
+    int option_index;
+
+    if (argc < 2) {
+        fprintf(stderr, "can not encode without option, please type 'yamitranscode -h' to help\n");
+        return false;
+    }
+
+    while ((opt = getopt_long_only(argc, argv, "W:H:b:f:c:s:i:o:N:h:", long_opts,&option_index)) != -1)
+    {
+        switch (opt) {
+        case 'h':
+        case '?':
+            print_help (argv[0]);
+            return false;
+        case 'i':
+            para.inputFileName = optarg;
+            break;
+        case 'o':
+            para.outputFileName = optarg;
+            break;
+        case 'W':
+            para.oWidth = atoi(optarg);
+            break;
+        case 'H':
+            para.oHeight = atoi(optarg);
+            break;
+        case 'b':
+            para.m_encParams.bitRate = atoi(optarg) * 1024;//kbps to bps
+            break;
+        case 'f':
+            para.m_encParams.fps = atoi(optarg);
+            break;
+        case 'c':
+            para.m_encParams.codec = optarg;
+            break;
+        case 's':
+            if (strlen(optarg) == 4)
+                para.fourcc = VA_FOURCC(optarg[0], optarg[1], optarg[2], optarg[3]);
+            break;
+        case 'N':
+            para.frameCount = atoi(optarg);
+            break;
+        case 0:
+             switch (option_index) {
+                case 1:
+                    para.m_encParams.initQp = atoi(optarg);
+                    break;
+                case 2:
+                    para.m_encParams.rcMode = string_to_rc_mode(optarg);
+                    break;
+            }
+        }
+    }
+
+    if (para.inputFileName.empty()) {
+        fprintf(stderr, "can not encode without input file\n");
+        return false;
+    }
+    if (para.outputFileName.empty())
+        para.outputFileName = "test.264";
+
+    if ((para.m_encParams.rcMode == RATE_CONTROL_CBR) && (para.m_encParams.bitRate <= 0)) {
+        fprintf(stderr, "please make sure bitrate is positive when CBR mode\n");
+        return false;
+    }
+
+    if (!strncmp(para.inputFileName.c_str(), "/dev/video", strlen("/dev/video")) && !para.frameCount)
+        para.frameCount = 50;
+
+    return true;
+}
+
+SharedPtr<VppInput> createInput(TranscodeParams& para, const SharedPtr<VADisplay>& display)
 {
     SharedPtr<VppInput> input(VppInput::create(para.inputFileName.c_str()));
     if (!input) {
@@ -70,7 +180,7 @@ SharedPtr<VppInput> createInput(YamiEncodeParam& para, const SharedPtr<VADisplay
     return VppInputAsync::create(input, 3);
 }
 
-SharedPtr<VppOutput> createOutput(YamiEncodeParam& para, const SharedPtr<VADisplay>& display)
+SharedPtr<VppOutput> createOutput(TranscodeParams& para, const SharedPtr<VADisplay>& display)
 {
 
     SharedPtr<VppOutput> output = VppOutput::create(para.outputFileName.c_str(), para.fourcc, para.oWidth, para.oHeight);
@@ -88,7 +198,7 @@ SharedPtr<VppOutput> createOutput(YamiEncodeParam& para, const SharedPtr<VADispl
         NativeDisplay nativeDisplay;
         nativeDisplay.type = NATIVE_DISPLAY_VA;
         nativeDisplay.handle = (intptr_t)*display;
-        if (!outputEncode->config(nativeDisplay)) {
+        if (!outputEncode->config(nativeDisplay, &para.m_encParams)) {
             ERROR("config ouput encode failed");
             output.reset();
         }
@@ -163,9 +273,9 @@ public:
             if(!m_output->output(dest))
                 break;
             count++;
+            fps.addFrame();
             if(count >= m_cmdParam.frameCount)
                 break;
-            fps.addFrame();
         }
         fps.log();
         return true;
@@ -184,7 +294,7 @@ private:
     SharedPtr<VppOutput> m_output;
     SharedPtr<FrameAllocator> m_allocator;
     SharedPtr<IVideoPostProcess> m_vpp;
-    YamiEncodeParam m_cmdParam;
+    TranscodeParams m_cmdParam;
 };
 
 int main(int argc, char** argv)
